add shared object validation on libs installation

+++ scripts/build_xc3sprog.sh
@@ -39,8 +39,11 @@ cmake -DCMAKE_TOOLCHAIN_FILE="${WD}/android-cmake/android.toolchain.cmake" \
 make -j${N_JOBS}
 rm -rf "${ANDROID_LIB_ROOT}/xc3sprog"
 make install/strip
-echo "-- Installing on Android app libs: libxc3loader.so ==> app/src/main/libs/armeabi-v7a"
-cp "${ANDROID_LIB_ROOT}/xc3sprog/lib/libxc3loader.so" "${WD}/app/src/main/libs/armeabi-v7a/"
-echo -n "-- Details: "
-echo `ls -Ggs --time-style=iso "${WD}/app/src/main/libs/armeabi-v7a/libxc3loader.so" | sed "s|${WD}/app/src/main/libs/armeabi-v7a/||g"`
+
+if [ -f "${ANDROID_LIB_ROOT}/xc3sprog/lib/libxc3loader.so" ]; then
+    echo "-- Installing on Android app libs: libxc3loader.so ==> app/src/main/libs/armeabi-v7a"
+    cp "${ANDROID_LIB_ROOT}/xc3sprog/lib/libxc3loader.so" "${WD}/app/src/main/libs/armeabi-v7a/"
+    echo -n "-- Details: "
+    echo `ls -Ggs --time-style=iso "${WD}/app/src/main/libs/armeabi-v7a/libxc3loader.so" | sed "s|${WD}/app/src/main/libs/armeabi-v7a/||g"`
+fi
 cd "${WD}"
